# Improve performance of building JarModuleFinder