--- a/src/main/java/cpw/mods/niofs/union/UnionFileSystem.java
+++ b/src/main/java/cpw/mods/niofs/union/UnionFileSystem.java
@@ -44,7 +44,7 @@ public class UnionFileSystem extends FileSystem {
         this.key = key;
         this.basepaths = IntStream.range(0, basepaths.length)
                 .mapToObj(i->basepaths[basepaths.length - i - 1])
-                .filter(Files::exists)
+                .filter(UnionFileSystem::exists)
                 .toList(); // we flip the list so later elements are first in search order.
         this.embeddedFileSystems = this.basepaths.stream().filter(path -> !Files.isDirectory(path))
                 .map(UnionFileSystem::openFileSystem)
@@ -146,7 +146,7 @@ public class UnionFileSystem extends FileSystem {
         return this.basepaths.stream()
                 .map(p->toRealPath(p , path))
                 .filter(p->p!=notExistingPath)
-                .filter(Files::exists)
+                .filter(UnionFileSystem::exists)
                 .findFirst();
     }
 
@@ -155,7 +155,7 @@ public class UnionFileSystem extends FileSystem {
                 .filter(p -> testFilter(toRealPath(p, path), p))
                 .map(p->toRealPath(p, path))
                 .filter(p->p!=notExistingPath)
-                .filter(Files::exists)
+                .filter(UnionFileSystem::exists)
                 .findFirst();
     }
 
@@ -173,7 +173,7 @@ public class UnionFileSystem extends FileSystem {
                 // We need to know the full path for the filter
                 Path realPath = toRealPath(base, path);
                 if (realPath != notExistingPath) {
-                    if (Files.exists(realPath)) {
+                    if (exists(realPath)) {
                         Optional<BasicFileAttributes> fileAttributes = this.getFileAttributes(realPath);
                         if (fileAttributes.isPresent()) {
                             if (testFilter(realPath, base)) {
@@ -187,6 +187,13 @@ public class UnionFileSystem extends FileSystem {
         } else {
             throw new UnsupportedOperationException();
         }
+    }
+
+    public static boolean exists(final Path p) {
+        if (p instanceof UnionPath unionPath) {
+            return unionPath.getFileSystem().findFirstFiltered(unionPath).isPresent();
+        }
+        return Files.exists(p);
     }
 
     public void checkAccess(final UnionPath p, final AccessMode... modes) throws IOException {
@@ -238,7 +245,7 @@ public class UnionFileSystem extends FileSystem {
         final var allpaths = new LinkedHashSet<Path>();
         for (final var bp : basepaths) {
             final var dir = toRealPath(bp, path);
-            if (dir == notExistingPath || !Files.exists(dir)) continue;
+            if (dir == notExistingPath || !exists(dir)) continue;
             final var isSimple = embeddedFileSystems.containsKey(bp);
             final var ds = Files.newDirectoryStream(dir, filter);
             StreamSupport.stream(ds.spliterator(), false)