The union FS is a bit slow currently.
The JarModuleFinder.of() call of the ModLauncher ModuleLayerHandler takes about 9 seconds on my machine in a forge dev environment, and most of this time is spend on getPackages, which iterates through the entire FileSystem.

By injecting the following code in JarModuleFinder.of(), the actual time needed can be profiled:
```java
        long start = System.currentTimeMillis();
        var moduleFinder = new JarModuleFinder(jars);
        long stop = System.currentTimeMillis();
        System.out.println("Creating " + JarModuleFinder.class.getSimpleName() + " for " + jars.length + " jars took " + (stop - start) + " ms!");
        return moduleFinder;
```
During startup in a development environment, this is called four times, and a quick VisualVM capture shows the longest run comes from `ModuleLayerHandler.buildLayer();`
Here are my timings without this patch:
```
Creating JarModuleFinder for 2 jars took 9263 ms!
Creating JarModuleFinder for 2 jars took 9351 ms!
Creating JarModuleFinder for 2 jars took 9243 ms!
```
I took the timings on my windows machine with the forge dev environment on my SSD.

Using VisualVM to identify hot spots, a few optimizations were made.
I tried to split them into indiviual commits, though some smaller ones are merged into one commit.
The improvments are considerable:
```
Creating JarModuleFinder for 2 jars took 2887 ms!
Creating JarModuleFinder for 2 jars took 2925 ms!
Creating JarModuleFinder for 2 jars took 2849 ms!
```
The time required to build the JarModuleFinder was reduced to about 1/3 of the original time required.