@LexManos 2021-07-26
Yes, the Union file system is one of the many things on my TODO to proper unit tests and performance tests for.
Any chance you can throw in some benchmark unit tests?