@ichttt 2021-07-27
RE: Some benchmarks:
I think it would make sense to introduce some sort of JMH benchmark in the securehjarhandler as well. I hope I'll find the time to do so in the next days