@ichttt 2021-07-30
I created a benchmark as a seperate PR, see #7 
Here are the numbers:
Base (without these changes): 16,144 +- 0,484  ms
This PR: 9,875 +- 0,586 ms

While the improvement is not as great as the improvement in the "real world" tests, it still shows a good improvement over the base,