@LexManos 2021-08-02
I've pulled your JMH PR, so be sure to rebase for that and make and more tests needed to highlight your specific changes.