@ichttt 2021-08-07
Sorry it took so long, but here are per-commit numbers: https://gist.github.com/ichttt/0ec28a8f5eb67f3cb24cb00e6cd9af8c
As you can see, 12110bc did not really have an impact on performance. Thinking a bit more about it, that makes sense, as the paths inside the UnionFileSystem are zip/real fs paths, and not union paths. A speedup coukld be achived if all calls to Files.exists outside of  the UnionFS called the custom impl, but that doesn't seem feasable, so I ended up reverting the commit.